move _Cache and rename Cache to CacheManager
