+++ tests/unit/utils/test_action.py
@@ -20,10 +20,7 @@ from unittest.mock import PropertyMock
 from juju.action import Action
 from juju.model import Model
 
-from juju_verify.utils.action import (
-    data_from_action,
-    cache
-)
+from juju_verify.utils.action import data_from_action
 
 
 def test_data_from_action(mocker):
@@ -43,34 +40,3 @@ def test_data_from_action(mocker):
 
     output = data_from_action(action, 'foo', default)
     assert output == default
-
-
-def test_cache():
-    """Test cache functions."""
-    default_cache_maxsize = cache.maxsize
-    key_1, action_1 = hash("test-1"), Action("1", Model())
-    key_2, action_2 = hash("test-2"), Action("2", Model())
-    key_3, action_3 = hash("test-3"), Action("3", Model())
-
-    cache.maxsize = 2  # change the maximum cache size for testing purposes
-    assert cache.maxsize == 2
-    cache[key_1] = action_1
-    cache[key_2] = action_2
-    assert key_1 in cache
-    assert key_2 in cache
-    assert key_3 not in cache
-    assert cache.keys == [key_1, key_2]  # check keys order
-    assert cache[key_1] == action_1
-    assert cache.keys == [key_2, key_1]  # check keys order
-    cache[key_3] = action_3
-    assert key_2 not in cache
-    assert key_3 in cache
-    assert cache.keys == [key_1, key_3]  # check keys order
-    cache.clear()
-    assert key_1 not in cache
-    assert key_2 not in cache
-    assert key_3 not in cache
-    # set the maximum cache size back to the default value
-    cache.maxsize = default_cache_maxsize
-    assert cache.maxsize == default_cache_maxsize
-    cache.clear()
