+++ tests/unit/utils/test_unit.py
@@ -25,7 +25,7 @@ from juju.unit import Unit
 from pytest import raises
 
 from juju_verify.exceptions import VerificationError, CharmException
-from juju_verify.utils.cache import action_cache
+from juju_verify.utils.action import cache_manager
 from juju_verify.utils.unit import (
     run_action_on_units,
     verify_charm_unit,
@@ -70,7 +70,7 @@ async def test_run_action():
     unit_2 = MagicMock()
     unit_1.entity_id.return_value = 1
     unit_2.entity_id.return_value = 2
-    action_cache.disable()  # disable run action cache usage
+    cache_manager.disable()  # disable run action cache usage
     unit_1.run_action.side_effect = unit_2.run_action.side_effect = mock_unit_run_action
 
     # test run_action once
@@ -94,7 +94,7 @@ async def test_run_action():
     unit_2.run_action.reset_mock()
 
     # test run_action multiple times with cache
-    action_cache.enable()  # disable run action cache usage
+    cache_manager.enable()  # disable run action cache usage
     await run_action(unit_1, "action-1", params=dict(format="json"))  # uses the cache
     await run_action(unit_1, "action-2")  # uses the cache
     await run_action(unit_1, "action-3")
